moved code related to snapshots to appropriate folder
