+++ src/Postgres/src/Eventuous.Postgresql/Eventuous.Postgresql.csproj
@@ -1,26 +1,26 @@
 <Project Sdk="Microsoft.NET.Sdk">
     <ItemGroup>
-        <ProjectReference Include="$(CoreRoot)\Eventuous.Subscriptions\Eventuous.Subscriptions.csproj"/>
-        <ProjectReference Include="$(CoreRoot)\Eventuous.Persistence\Eventuous.Persistence.csproj"/>
-        <ProjectReference Include="$(SrcRoot)\Relational\src\Eventuous.Sql.Base\Eventuous.Sql.Base.csproj"/>
-        <ProjectReference Include="$(CoreRoot)\Eventuous.Producers\Eventuous.Producers.csproj"/>
+        <ProjectReference Include="$(CoreRoot)\Eventuous.Subscriptions\Eventuous.Subscriptions.csproj" />
+        <ProjectReference Include="$(CoreRoot)\Eventuous.Persistence\Eventuous.Persistence.csproj" />
+        <ProjectReference Include="$(SrcRoot)\Relational\src\Eventuous.Sql.Base\Eventuous.Sql.Base.csproj" />
+        <ProjectReference Include="$(CoreRoot)\Eventuous.Producers\Eventuous.Producers.csproj" />
     </ItemGroup>
     <ItemGroup>
-        <PackageReference Include="Microsoft.Extensions.Options.ConfigurationExtensions"/>
-        <PackageReference Include="Npgsql"/>
-        <PackageReference Include="Npgsql.DependencyInjection"/>
+        <PackageReference Include="Microsoft.Extensions.Options.ConfigurationExtensions" />
+        <PackageReference Include="Npgsql" />
+        <PackageReference Include="Npgsql.DependencyInjection" />
     </ItemGroup>
     <ItemGroup>
-        <EmbeddedResource Include="Scripts\1_Schema.sql"/>
-        <EmbeddedResource Include="Scripts\2_AppendEvents.sql"/>
-        <EmbeddedResource Include="Scripts\3_CheckStream.sql"/>
-        <EmbeddedResource Include="Scripts\4_ReadAllForwards.sql"/>
-        <EmbeddedResource Include="Scripts\5_ReadStreamBackwards.sql"/>
-        <EmbeddedResource Include="Scripts\6_ReadStreamForwards.sql"/>
-        <EmbeddedResource Include="Scripts\7_ReadStreamSub.sql"/>
-        <EmbeddedResource Include="Scripts\8_TruncateStream.sql"/>
-        <EmbeddedResource Include="Scripts\9_TryInsertTombstone.sql"/>
-        <EmbeddedResource Include="SnapshotScripts\1_SnapshotSchema.sql"/>
+        <EmbeddedResource Include="Scripts\1_Schema.sql" />
+        <EmbeddedResource Include="Scripts\2_AppendEvents.sql" />
+        <EmbeddedResource Include="Scripts\3_CheckStream.sql" />
+        <EmbeddedResource Include="Scripts\4_ReadAllForwards.sql" />
+        <EmbeddedResource Include="Scripts\5_ReadStreamBackwards.sql" />
+        <EmbeddedResource Include="Scripts\6_ReadStreamForwards.sql" />
+        <EmbeddedResource Include="Scripts\7_ReadStreamSub.sql" />
+        <EmbeddedResource Include="Scripts\8_TruncateStream.sql" />
+        <EmbeddedResource Include="Scripts\9_TryInsertTombstone.sql" />
+        <EmbeddedResource Include="Snapshots\Scripts\1_SnapshotSchema.sql" />
     </ItemGroup>
     <ItemGroup>
         <Compile Include="$(CoreRoot)\Eventuous.Shared\Tools\TaskExtensions.cs">
@@ -31,11 +31,11 @@
         </Compile>
     </ItemGroup>
     <ItemGroup>
-        <InternalsVisibleTo Include="Eventuous.Tests.Postgres"/>
+        <InternalsVisibleTo Include="Eventuous.Tests.Postgres" />
     </ItemGroup>
     <ItemGroup>
-        <Using Include="Npgsql"/>
-        <Using Include="NpgsqlTypes"/>
-        <Using Include="Eventuous.Tools"/>
+        <Using Include="Npgsql" />
+        <Using Include="NpgsqlTypes" />
+        <Using Include="Eventuous.Tools" />
     </ItemGroup>
 </Project>
