+++ src/Postgres/src/Eventuous.Postgresql/Schema.cs
@@ -36,7 +36,7 @@ public class Schema(string schema = Schema.DefaultSchema) {
     public async Task CreateSchema(NpgsqlDataSource dataSource, ILogger<Schema>? log, CancellationToken cancellationToken = default) {
         log?.LogInformation("Creating schema {Schema}", schema);
         var names = Assembly.GetManifestResourceNames()
-            .Where(x => x.EndsWith(".sql") && !x.Contains("SnapshotScripts"))
+            .Where(x => x.EndsWith(".sql") && !x.Contains("Snapshots.Scripts"))
             .OrderBy(x => x);
 
         await using var connection = await dataSource.OpenConnectionAsync(cancellationToken).NoContext();
