+++ src/Postgres/src/Eventuous.Postgresql/Snapshots/Scripts/1_SnapshotSchema.sql
